Remove classes in different files
